Move GetDockerVersion to util() to avoid import cycles
